Remove variables table code
