+++ src/datadoc/frontend/fields/display_variables.py
@@ -2,12 +2,15 @@
 
 from __future__ import annotations
 
-import typing
+import functools
 from enum import Enum
 
-from datadoc_model import model
-
-from datadoc.frontend.fields.display_base import DisplayVariablesMetadata
+from datadoc import enums
+from datadoc.frontend.fields.display_base import VariablesCheckboxField
+from datadoc.frontend.fields.display_base import VariablesDropdownField
+from datadoc.frontend.fields.display_base import VariablesInputField
+from datadoc.frontend.fields.display_base import get_enum_options_for_language
+from datadoc.frontend.fields.display_base import get_multi_language_metadata
 
 
 class VariableIdentifiers(str, Enum):
@@ -33,117 +36,138 @@ class VariableIdentifiers(str, Enum):
     CONTAINS_DATA_UNTIL = "contains_data_until"
 
 
-DISPLAY_VARIABLES = {
-    VariableIdentifiers.SHORT_NAME: DisplayVariablesMetadata(
+DISPLAY_VARIABLES: dict[
+    VariableIdentifiers,
+    VariablesInputField | VariablesDropdownField | VariablesCheckboxField,
+] = {
+    VariableIdentifiers.SHORT_NAME: VariablesInputField(
         identifier=VariableIdentifiers.SHORT_NAME.value,
         display_name="Kortnavn",
         description="Fysisk navn på variabelen i datasettet. Bør tilsvare anbefalt kortnavn.",
         obligatory=True,
         editable=False,
     ),
-    VariableIdentifiers.NAME: DisplayVariablesMetadata(
+    VariableIdentifiers.NAME: VariablesInputField(
         identifier=VariableIdentifiers.NAME.value,
         display_name="Navn",
         description="Variabelnavn kan arves fra VarDef, men kan også dokumenteres/endres her.",
         obligatory=True,
         multiple_language_support=True,
+        type="text",
     ),
-    VariableIdentifiers.DATA_TYPE: DisplayVariablesMetadata(
+    VariableIdentifiers.DATA_TYPE: VariablesDropdownField(
         identifier=VariableIdentifiers.DATA_TYPE.value,
         display_name="Datatype",
         description="Datatype",
         obligatory=True,
-        presentation="dropdown",
+        options_getter=functools.partial(
+            get_enum_options_for_language,
+            enums.DataType,
+        ),
     ),
-    VariableIdentifiers.VARIABLE_ROLE: DisplayVariablesMetadata(
+    VariableIdentifiers.VARIABLE_ROLE: VariablesDropdownField(
         identifier=VariableIdentifiers.VARIABLE_ROLE.value,
         display_name="Variabelens rolle",
         description="Variabelens rolle i datasett",
         obligatory=True,
-        presentation="dropdown",
+        options_getter=functools.partial(
+            get_enum_options_for_language,
+            enums.VariableRole,
+        ),
     ),
-    VariableIdentifiers.DEFINITION_URI: DisplayVariablesMetadata(
+    VariableIdentifiers.DEFINITION_URI: VariablesInputField(
         identifier=VariableIdentifiers.DEFINITION_URI.value,
         display_name="Definition URI",
         description="En lenke (URI) til variabelens definisjon i SSB (Vardok/VarDef)",
         url=True,
         obligatory=True,
+        type="url",
     ),
-    VariableIdentifiers.DIRECT_PERSON_IDENTIFYING: DisplayVariablesMetadata(
+    VariableIdentifiers.DIRECT_PERSON_IDENTIFYING: VariablesCheckboxField(
         identifier=VariableIdentifiers.DIRECT_PERSON_IDENTIFYING.value,
-        display_name="DPI",
+        display_name="Direkte personidentifiserende informasjon",
         description="Direkte personidentifiserende informasjon (DPI)",
         obligatory=True,
-        presentation="dropdown",
     ),
-    VariableIdentifiers.DATA_SOURCE: DisplayVariablesMetadata(
+    VariableIdentifiers.DATA_SOURCE: VariablesInputField(
         identifier=VariableIdentifiers.DATA_SOURCE.value,
         display_name="Datakilde",
         description="Datakilde. Settes på datasettnivå, men kan overstyres på variabelforekomstnivå.",
         multiple_language_support=True,
+        type="text",
     ),
-    VariableIdentifiers.POPULATION_DESCRIPTION: DisplayVariablesMetadata(
+    VariableIdentifiers.POPULATION_DESCRIPTION: VariablesInputField(
         identifier=VariableIdentifiers.POPULATION_DESCRIPTION.value,
         display_name="Populasjonen",
         description="Populasjonen variabelen beskriver kan spesifiseres nærmere her. Settes på datasettnivå, men kan overstyres på variabelforekomstnivå.",
         multiple_language_support=True,
+        type="text",
     ),
-    VariableIdentifiers.COMMENT: DisplayVariablesMetadata(
+    VariableIdentifiers.COMMENT: VariablesInputField(
         identifier=VariableIdentifiers.COMMENT.value,
         display_name="Kommentar",
         description="Ytterligere presiseringer av variabeldefinisjon",
         multiple_language_support=True,
+        type="text",
     ),
-    VariableIdentifiers.TEMPORALITY_TYPE: DisplayVariablesMetadata(
+    VariableIdentifiers.TEMPORALITY_TYPE: VariablesDropdownField(
         identifier=VariableIdentifiers.TEMPORALITY_TYPE.value,
         display_name="Temporalitetstype",
         description="Temporalitetstype. Settes enten for variabelforekomst eller datasett. Se Temporalitet, hendelser og forløp.",
-        presentation="dropdown",
+        options_getter=functools.partial(
+            get_enum_options_for_language,
+            enums.TemporalityTypeType,
+        ),
     ),
-    VariableIdentifiers.MEASUREMENT_UNIT: DisplayVariablesMetadata(
+    VariableIdentifiers.MEASUREMENT_UNIT: VariablesInputField(
         identifier=VariableIdentifiers.MEASUREMENT_UNIT.value,
         display_name="Måleenhet",
         description="Måleenhet. Eksempel: NOK eller USD for valuta, KG eller TONN for vekt. Se også forslag til SSBs måletyper/måleenheter.",
+        type="text",
     ),
-    VariableIdentifiers.FORMAT: DisplayVariablesMetadata(
+    VariableIdentifiers.FORMAT: VariablesInputField(
         identifier=VariableIdentifiers.FORMAT.value,
         display_name="Format",
         description="Verdienes format (fysisk format eller regulært uttrykk) i maskinlesbar form ifm validering. Dette kan benyttes som en ytterligere presisering av datatypen (dataType) i de tilfellene hvor dette er relevant. ",
     ),
-    VariableIdentifiers.CLASSIFICATION_URI: DisplayVariablesMetadata(
+    VariableIdentifiers.CLASSIFICATION_URI: VariablesInputField(
         identifier=VariableIdentifiers.CLASSIFICATION_URI.value,
         display_name="Kodeverkets URI",
         description="Lenke (URI) til gyldige kodeverk (klassifikasjon eller kodeliste) i KLASS",
         url=True,
+        type="url",
     ),
-    VariableIdentifiers.SENTINEL_VALUE_URI: DisplayVariablesMetadata(
+    VariableIdentifiers.SENTINEL_VALUE_URI: VariablesInputField(
         identifier=VariableIdentifiers.SENTINEL_VALUE_URI.value,
         display_name="Spesialverdienes URI",
         description="En lenke (URI) til en oversikt over 'spesialverdier' som inngår i variabelen.",
         url=True,
+        type="url",
     ),
-    VariableIdentifiers.INVALID_VALUE_DESCRIPTION: DisplayVariablesMetadata(
+    VariableIdentifiers.INVALID_VALUE_DESCRIPTION: VariablesInputField(
         identifier=VariableIdentifiers.INVALID_VALUE_DESCRIPTION.value,
         display_name="Ugyldige verdier",
         description="En beskrivelse av ugyldige verdier som inngår i variabelen dersom spesialverdiene ikke er tilstrekkelige eller ikke kan benyttes.",
         multiple_language_support=True,
     ),
-    VariableIdentifiers.IDENTIFIER: DisplayVariablesMetadata(
+    VariableIdentifiers.IDENTIFIER: VariablesInputField(
         identifier=VariableIdentifiers.IDENTIFIER.value,
         display_name="Unik ID",
         description="Unik SSB identifikator for variabelforekomsten i datasettet",
         obligatory=False,
         editable=False,
     ),
-    VariableIdentifiers.CONTAINS_DATA_FROM: DisplayVariablesMetadata(
+    VariableIdentifiers.CONTAINS_DATA_FROM: VariablesInputField(
         identifier=VariableIdentifiers.CONTAINS_DATA_FROM.value,
         display_name="Inneholder data f.o.m.",
         description="Variabelforekomsten i datasettet inneholder data fra og med denne dato.",
+        type="date",
     ),
-    VariableIdentifiers.CONTAINS_DATA_UNTIL: DisplayVariablesMetadata(
+    VariableIdentifiers.CONTAINS_DATA_UNTIL: VariablesInputField(
         identifier=VariableIdentifiers.CONTAINS_DATA_UNTIL.value,
         display_name="Inneholder data t.o.m.",
         description="Variabelforekomsten i datasettet inneholder data til og med denne dato.",
+        type="date",
     ),
 }
 
@@ -151,22 +175,14 @@ MULTIPLE_LANGUAGE_VARIABLES_METADATA = [
     m.identifier for m in DISPLAY_VARIABLES.values() if m.multiple_language_support
 ]
 
-URL_VARIABLES_METADATA = [m.identifier for m in DISPLAY_VARIABLES.values() if m.url]
-
-DISPLAYED_DROPDOWN_VARIABLES_METADATA = [
-    m.identifier for m in DISPLAY_VARIABLES.values() if m.presentation == "dropdown"
-]
-
-DISPLAYED_DROPDOWN_VARIABLES_TYPES = []
-
-types = typing.get_type_hints(model.Variable)
-
-for m in DISPLAY_VARIABLES.values():
-    if m.presentation == "dropdown":
-        field_type = typing.get_args(types[m.identifier])[0]
-        DISPLAYED_DROPDOWN_VARIABLES_TYPES.append(field_type)
-
+for v in DISPLAY_VARIABLES.values():
+    if v.multiple_language_support:
+        v.value_getter = get_multi_language_metadata
 
 OBLIGATORY_VARIABLES_METADATA = [
-    m.identifier for m in DISPLAY_VARIABLES.values() if m.obligatory and m.editable
+    m for m in DISPLAY_VARIABLES.values() if m.obligatory and m.editable
+]
+
+OPTIONAL_VARIABLES_METADATA = [
+    m for m in DISPLAY_VARIABLES.values() if not m.obligatory and m.editable
 ]
